backend/requirements.txt
Django==4.2.16
python-dotenv==1.0.1
djangorestframework==3.15.2
djoser==2.2.3
psycopg2-binary==2.9.9
flake8==7.1.1
isort==5.13.2
pillow==10.4.0
django-cors-headers==4.4.0
django-filter==23.1
hashids==1.3.1